# schema/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestRequest": {
        "additionalProperties": false,
        "description": "Backtest request.",
        "properties": {
          "bar_spec": {
            "description": "Bar specification as a string (e.g. \"1d\").",
            "type": [
              "string",
              "null"
            ]
          },
          "end": {
            "description": "Exclusive end date/time.",
            "type": [
              "string",
              "null"
            ]
          },
          "initial_capital": {
            "description": "Initial capital.",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "seed": {
            "description": "Seed, for reproducibility. Same seed and data give a byte-identical journal.",
            "format": "uint64",
            "minimum": 0.0,
            "type": [
              "integer",
              "null"
            ]
          },
          "start": {
            "description": "Inclusive start date/time.",
            "type": [
              "string",
              "null"
            ]
          },
          "strategy": {
            "description": "Strategy name or manifest id, as returned by `POST /strategies`.",
            "type": [
              "string",
              "null"
            ]
          },
          "universe": {
            "description": "Universe of instruments.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "BacktestResponse": {
        "description": "Backtest response.",
        "properties": {
          "assumptions": {
            "description": "What the run did *not* model, stated explicitly."
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestMetrics"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present once `status` is `completed`."
          },
          "run_id": {
            "description": "Server-assigned run id.",
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RunStatus"
              }
            ],
            "description": "Current lifecycle state."
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "type": "object"
      },
...
      "TradesResponse": {
        "description": "Trades response.",
        "properties": {
          "trades": {
            "description": "Realized trades.",
            "items": {
              "$ref": "#/components/schemas/Trade"
            },
            "type": "array"
          }
        },
        "required": [
          "trades"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated by honba-codegen from the Rust source of truth. DO NOT EDIT: regenerate with `make openapi`.",
    "title": "Honba API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/backtests": {
      "post": {
        "operationId": "post_backtests",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST backtests",
        "tags": [
          "runs"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/backtests/{id}": {
      "get": {
        "operationId": "get_backtests_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET backtests  id",
        "tags": [
          "runs"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/backtests/{id}/journal": {
      "get": {
        "operationId": "get_backtests_by_id_journal",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET backtests  id  journal",
        "tags": [
          "runs"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/bars/{id}": {
      "get": {
        "operationId": "get_bars_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "description": "Inclusive start time, RFC3339 or ISO-8601 date.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "tf",
            "required": false,
            "schema": {
              "description": "Timeframe/bar specification (e.g. \"1m\", \"1d\").",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "description": "Exclusive end time, RFC3339 or ISO-8601 date.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET bars  id",
        "tags": [
          "bars"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/capabilities": {
      "get": {
        "operationId": "get_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET capabilities",
        "tags": [
          "meta"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/depth/{id}": {
      "get": {
        "operationId": "get_depth_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "description": "Number of levels per side.",
              "format": "uint32",
              "minimum": 0.0,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepthResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET depth  id",
        "tags": [
          "depth"
        ],
        "x-honba-access": "read_only"
      }
    },
...
    "/instruments": {
      "get": {
        "operationId": "get_instruments",
        "parameters": [
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "description": "Exchange filter.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "description": "Symbol filter.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET instruments",
        "tags": [
          "instruments"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/instruments/{id}": {
      "get": {
        "operationId": "get_instruments_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET instruments  id",
        "tags": [
          "instruments"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/journals/{id}": {
      "get": {
        "operationId": "get_journals_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET journals  id",
        "tags": [
          "meta"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/orders": {
      "get": {
        "operationId": "get_orders",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET orders",
        "tags": [
          "trading"
        ],
        "x-honba-access": "read_only"
      },
      "post": {
        "operationId": "post_orders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST orders",
        "tags": [
          "trading"
        ],
        "x-honba-access": "write"
      }
    },
    "/orders/{id}": {
      "delete": {
        "operationId": "delete_orders_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "DELETE orders  id",
        "tags": [
          "trading"
        ],
        "x-honba-access": "write"
      }
    },
    "/positions/close": {
      "post": {
        "operationId": "post_positions_close",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST positions close",
        "tags": [
          "trading"
        ],
        "x-honba-access": "write"
      }
    },
    "/quotes": {
      "get": {
        "operationId": "get_quotes",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "description": "Return the latest quote known at this time (RFC3339 or ISO-8601 date, inclusive); the latest held when omitted.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "symbols",
            "required": false,
            "schema": {
              "description": "Comma-separated symbols.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "venue",
            "required": false,
            "schema": {
              "description": "Venue/exchange.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET quotes",
        "tags": [
          "quotes"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/schema": {
      "get": {
        "operationId": "get_schema",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET schema",
        "tags": [
          "meta"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/screener/scan": {
      "get": {
        "operationId": "get_screener_scan",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET screener scan",
        "tags": [
          "screener"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/strategies": {
      "get": {
        "operationId": "get_strategies",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategiesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET strategies",
        "tags": [
          "strategies"
        ],
        "x-honba-access": "read_only"
      },
      "post": {
        "operationId": "post_strategies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategiesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST strategies",
        "tags": [
          "strategies"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/strategies/verify": {
      "post": {
        "operationId": "post_strategies_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyIr"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST strategies verify",
        "tags": [
          "strategies"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/sweeps": {
      "post": {
        "operationId": "post_sweeps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SweepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "POST sweeps",
        "tags": [
          "runs"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/sweeps/{id}": {
      "get": {
        "operationId": "get_sweeps_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET sweeps  id",
        "tags": [
          "runs"
        ],
        "x-honba-access": "read_only"
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}